requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-crispy-forms==1.8.1
django-heroku==0.3.1
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
